requirements.txt
requests==2.18.4
numpy==1.14.3
scipy==1.1.0
redis==3.2.1
pandas==0.23.0
glog==0.3.1
flask==1.0.2
Pillow==5.1.0
scikit-learn==0.19.1
